2DPlatformerUpdate
BetterJump/ScrollCamA set of animations for the player, including Idle and Walking, at the very least
Only allow the player to jump when on the ground
A health bar/health indicator (text showing something like "health: 100" is sufficient here)
Hazards that damage the player on contact
A life system: if the player runs out of health, the player should lose a life and return to the starting location; if no lives remain, the game should restart.
Better Jumping: Many platformers include a variable jump height that makes jumping feel better (hold the jump button to jump higher); add this feature, as well as a double jump, that will allow the player to jump one time while in the air. You can refer to this video for better jumping:
Scrolling Camera: Include a camera that smoothly follows your player. Ensure that your level is long enough to show this off.